core accounts, modules, socket, and store refactors
• Add accounts.ts to handle account lifecycle, connection/reconnection, pings and login
• Refactor accounts_config.js to use shared storage and update account existence checks
• Update core.js to import/re-export new account, module, socket, and store functionality
• Remove legacy AccountsAddEdit modal and update module import paths
• Introduce new modules.ts for module registration and communication handling
• Create socket.ts for handling WebSocket messages and command sending
• Add stores.ts and types.ts to centralize Svelte stores and type definitions
• Update utils.ts and vitest.shims.d.ts for minor adjustments and global declarations
