+++ vitest.shims.d.ts
@@ -1,1 +1,5 @@
 /// <reference types="@vitest/browser/providers/playwright" />
+
+declare const __BUILD_DATE__: string;
+declare const __COMMIT_HASH__: string;
+declare const __BRANCH__: string;
